requirements.txt
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.1.0
django-filter==23.4
django-extensions==3.2.3
django-environ==0.11.2
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-storages==1.14.2
boto3==1.34.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
markdown==3.5.1
django-taggit==4.0.0
